fix: migrated to pages directory, feat: added export script
