--- src/app/head.tsx
@@ -1,9 +1,0 @@
-export default function Head() {
-  return (
-    <>
-      <title>Ask Buddie FAQ</title>
-      <meta name="viewport" content="width=device-width, initial-scale=1" />
-      <link rel="icon" href="/favicon.ico" />
-    </>
-  );
-}
