--- src/app/layout.tsx
@@ -1,14 +1,0 @@
-import "../styles/globals.css";
-
-type Props = {
-  children: React.ReactNode;
-};
-
-export default function RootLayout({ children }: Props) {
-  return (
-    <html lang="en">
-      <head />
-      <body>{children}</body>
-    </html>
-  );
-}
