+++ src/pages/_app.tsx
@@ -1,0 +1,6 @@
+import { AppProps } from "next/app";
+import "../styles/globals.css";
+
+export default function MyApp({ Component, pageProps }: AppProps) {
+  return <Component {...pageProps} />;
+}
