+++ src/styles/globals.css
@@ -1,5 +1,3 @@
-/* ./app/global.css */
-
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
